+++ features/bootstrap/Template/i18n/pt_PT/default.php
@@ -1,0 +1,14 @@
+<?php
+
+/**
+ * Messages translation
+ */
+
+return [
+    '' => array(
+        'plural_forms' => 'nplurals=2; plural=n!=1;'
+    ),
+    'Hello world' => 'Olá mundo',
+    'User' => ['Utilizador', 'Utilizadores'],
+    'Users' => ''
+];