+++ src/Engine/AbstractEngine.php
@@ -10,7 +10,7 @@
 namespace Slick\Template\Engine;
 
 use Slick\Common\Base;
-use Slick\Template\EngineInterface;
+use Slick\Template\TemplateEngineInterface;
 
 /**
  * AbstractEngine
@@ -18,12 +18,12 @@ use Slick\Template\EngineInterface;
  * @package   Slick\Template\Engine
  * @author    Filipe Silva <silvam.filipe@gmail.com>
  */
-abstract class AbstractEngine extends Base implements EngineInterface
+abstract class AbstractEngine extends Base implements TemplateEngineInterface
 {
     /**
      * Handles the initialization if engine is already initialized
      *
-     * @return AbstractEngine
+     * @return AbstractEngine|self|$this
      */
     public function initialize()
     {