+++ CHANGELOG.md
@@ -2,6 +2,21 @@
 
 All Notable changes to `Slick/template` will be documented in this file.
 
+## 1.2.0 - 2015-02-10
+
+### Added
+- Support for extensions
+- I18n extension
+- Text extension
+
+### Removed
+- Twig extension
+
+### Deprecated
+- EngineInterface is now deprecated and should be replaced by
+  TemplateEngineInterface for a more accurate meaning to the interface name.
+
+
 ## 1.1.0 - 2015-04-14
 
 ### Added